Added Texture Atlas Functions
Before there was only one texture file, now there's two texture atlases.

Created a function that returns the texture UV location based on a selection on a atlas grid (e.g., 0,0 to select the first item on the atlas).

Created another function that takes in the selection mentioned before and allows the programmer to input the UV locations into the shader.