Fixed deltaTime bug + updated README.md instructions
Before, pressing the LeftShift and Space key went up and down (the y axis), but on other computers, this would be too slow (not matched to time). Now every computer moves the same speed.

I wrote the wrong instructions, this project needs my "GLFW" #include folders instead of "LIBRARIES" (which was from another project)

Please include the "GLFW" folder on the C:/ drive in order to run this project.