Improved main menu text handling
In my text render I use \\ to line break. But instead of adding onto a string with outputText += "\\string" by the time I got to the 5th string it would look like outputText += "\\\\\\\\\\string".

All I did was add an outputText variable so I could just add text onto it so it could be expanded easier.